upload-service/requirements.txt
Flask==3.1.0
google-cloud-storage==2.18.0
google-cloud-tasks==2.16.4
gunicorn==23.0.0
